VERSION APP EXTRACCION 1.7 Corrijo algunas llamadas de rango afip segun necesite que me traiga el rango anterior o el rango donde esta la fecha Agrego metodo ControlarFechaFile para usar el archivo fileAfip.properties para guardar el ultimo rango extraido. Quito el borrado del archivo afip que estaba ya que no lo debo borrar. sino controlar que no exista

+++ src/main/java/ar/com/bambu/communicator/reply/hassar/ReporteElectronico.java
@@ -4,6 +4,7 @@ import ar.com.bambu.jpos.HassarFrameMsg;
 import ar.com.bambu.utils.Ascii85;
 
 import java.io.*;
+import java.util.Properties;
 
 public class ReporteElectronico extends AbstractReply {
 
@@ -49,15 +50,59 @@ public class ReporteElectronico extends AbstractReply {
             throw new IOException("File not ready yet to be saved");
         }
         File file = new File(FILE_NAME + "_" + nroPuntoVta + "_" + rangoI + "_a_" + rangoF);
-        file.delete();
+        /*file.delete();
         File debug = new File("debug.txt");
-        debug.delete();
+        debug.delete();*/
 
-        String asci85 = this.data.replace("<~", "");
-        asci85 = asci85.replace("~>", "");
-        OutputStream os = new FileOutputStream(FILE_NAME + "_" + nroPuntoVta + "_" + rangoI + "_a_" + rangoF + ".zip");
-        os.write(Ascii85.decode(asci85));
-        os.close();
+        //veamo si tenemos una copia del archivo ya creada asi no la grabamos de nuevo
+        String sCarpAct = System.getProperty("user.dir");
+        File carpeta = new File(sCarpAct);
+
+        String fileBackup = carpeta.toString() +  "/backupAfip/" + FILE_NAME + "_" + nroPuntoVta + "_" + rangoI + "_a_" + rangoF + ".zip";
+        File archivo= new File(fileBackup);
+
+        String fileBackup2 = carpeta.toString() +  "/" + FILE_NAME + "_" + nroPuntoVta + "_" + rangoI + "_a_" + rangoF + ".zip";
+        File archivo2= new File(fileBackup2);
+        try{
+            if (archivo.exists() || archivo2.exists()) {
+                System.out.println("OJO: ¡¡YA  existe ESTE ARCHIVO DE REPORTE");
+                throw new Exception ("YA  existe ESTE ARCHIVO DE REPORTE");
+            }
+
+            String asci85 = this.data.replace("<~", "");
+            asci85 = asci85.replace("~>", "");
+            OutputStream os = new FileOutputStream(FILE_NAME + "_" + nroPuntoVta + "_" + rangoI + "_a_" + rangoF + ".zip");
+            os.write(Ascii85.decode(asci85));
+            os.close();
+            //guardemos el rango en el archivio
+            // Get the file
+            File f = new File("fileAfip.properties");
+
+            // Check if the specified file
+            // Exists or not
+            if (!f.exists()) {
+                f.createNewFile();
+            }
+
+
+        }
+        catch(Exception e){
+            System.out.println(e);
+        } finally {
+            File f = new File("fileAfip.properties");
+            Properties prop = new Properties();
+            // Leer el archivo de propiedades db.properties
+            /// Guarde las propiedades en el archivo new.properties
+            FileOutputStream oFile = new FileOutputStream(f, false);// verdadero significa abierto adicional
+            prop.setProperty("PtoVta", String.valueOf(nroPuntoVta));
+            prop.setProperty("RANGOI", String.valueOf(rangoI));
+            prop.setProperty("RANGOF", String.valueOf(rangoF));
+            prop.store(oFile, "");
+            oFile.close();
+
+        }
+
+
     }
 
     @Override
